Move trace id middleware to log package
